Reorganize packages
This commit moves the types package to src/, instead on having it
as a package under storage.

Also, a stack implementation was added to help one of the created tests.

Another validation was added to the LottoInputDTO to check whether
a number is being used as fixed and most sorted at the same time.

Some variables that where in snake case were renamed to use camel case.

Last, but not least, the LottoInputDTO validations were moved from
router to types package.